rename caregiver view
