+++ mobile/StaffApp.tsx
@@ -6,7 +6,7 @@ import Activity from './views/staff/Activity'
 import Client from './views/staff/Client'
 import Staff from './views/staff/Staff'
 import ClientProfile from './views/staff/ClientProfile'
-import CaregiverList from './views/staff/CaregiverList'
+import Caregiver from './views/staff/Caregiver'
 
 const Stack = createNativeStackNavigator()
 
@@ -19,7 +19,7 @@ export default function StaffApp() {
         <Stack.Screen name="Client" component={Client} />
         <Stack.Screen name="Staff" component={Staff} />
         <Stack.Screen name="ClientProfile" component={ClientProfile} />
-        <Stack.Screen name="CaregiverList" component={CaregiverList} />
+        <Stack.Screen name="Caregiver" component={Caregiver} />
       </Stack.Navigator>
     </NavigationContainer>
   )
